pyproject.toml
[project]
name = "one-geosite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pylint>=3.3.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
